requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
paho-mqtt==1.6.1
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
